requirements.txt
tf-keras
python-dotenv
transformers
pillow
langchain
langchain-groq
scipy
bark
tensorflow